WH_지도 좌표 데이터를 통해서 상세페이지에서 글에 해당하는 지도보여주기 구현함.

+++ SixQuense_project/src/main/java/com/web/accompany/model/dto/Coordinate.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class Coordinate {
 	private int coordinateNo;
 	private String nation;
-	private int latitude;
-	private int longitude;
+	private double latitude;
+	private double longitude;
 	private Continent continent; 
 }
